Finish evaluating Dual Attention Block

+++ preprocessing/dataset_class.py
@@ -28,7 +28,7 @@ class MSIDataset(Dataset):
         image = image.convert('RGB')
         image_to_tensor = self.to_tensor(image)
         int_label = np.array(self.int_labels[index])
-        int_label = torch.from_numpy(int_label)
+        int_label = torch.from_numpy(int_label).float()
         return image_to_tensor, int_label
     
     def __len__(self):
